+++ MonkeyBot/Modules/Feeds/IFeedService.cs
