+++ MonkeyBot/Modules/GameServers/BaseGameServerService.cs
