+++ MonkeyBot/Modules/GameServers/IGameServerService.cs
