+++ MonkeyBot/Modules/GameServers/MineCraft/MineQuery.cs
