+++ MonkeyBot/Modules/GameServers/MineCraft/MineQueryResult.cs
