+++ MonkeyBot/Modules/Guild/GuildService.cs
