+++ MonkeyBot/Modules/Guild/IGuildService.cs
