+++ MonkeyBot/Modules/Guild/ModeratorModule.cs
