+++ MonkeyBot/Modules/Misc/MiscModule.cs
@@ -2,7 +2,6 @@
 using DSharpPlus.CommandsNext.Attributes;
 using DSharpPlus.Entities;
 using MonkeyBot.Common;
-using MonkeyBot.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,34 +12,6 @@ namespace MonkeyBot.Modules
 {
     public class MiscModule : BaseCommandModule
     {
-        //private readonly MonkeyDBContext dbContext;
-        private readonly IGuildService guildService;
-
-        public MiscModule(IGuildService guildService)
-        {
-            this.guildService = guildService;
-        }
-
-        [Command("Rules")]
-        [Description("The bot replies with the server rules")]
-        [RequireGuild]
-        public async Task ListRulesAsync(CommandContext ctx)
-        {
-            List<string> rules = (await guildService.GetOrCreateConfigAsync(ctx.Guild.Id).ConfigureAwait(false)).Rules;
-            if (rules == null || rules.Count < 1)
-            {
-                _ = await ctx.RespondAsync("No rules set!").ConfigureAwait(false);
-                return;
-            }
-            var builder = new DiscordEmbedBuilder()
-                .WithColor(DiscordColor.DarkGreen)
-                .WithTitle($"Rules of {ctx.Guild.Name}")
-                .WithDescription(string.Join(Environment.NewLine, rules));
-
-            _ = await ctx.RespondDeletableAsync(embed: builder.Build()).ConfigureAwait(false);
-
-        }
-
         [Command("FindMessageID")]
         [Description("Gets the message id of a message in the current channel with the provided message text")]
         [RequireGuild]
