+++ MonkeyBot/Modules/Misc/PollModule.cs
