+++ MonkeyBot/Modules/Trivia/ITriviaQuestion.cs
