+++ MonkeyBot/Modules/Trivia/ITriviaService.cs
