+++ MonkeyBot/Modules/Trivia/OTDBDifficultyConverter.cs
