+++ MonkeyBot/Modules/Trivia/OTDBQuestionTypeConverter.cs
