+++ MonkeyBot/Modules/Trivia/OTDBResponse.cs
