+++ MonkeyBot/Modules/Trivia/TriviaStatus.cs
