+++ MonkeyBot/Services/BattlefieldNewsService.cs
@@ -4,7 +4,6 @@ using HtmlAgilityPack;
 using Microsoft.Extensions.Logging;
 using MonkeyBot.Models;
 using System;
-using System.ComponentModel;
 using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
